app/requirements.txt
Flask
psycopg2-binary
pytest
requests

